[UPD] - extract config, validation, and encapsulate state;
